Login form validation implemented
