--- AppTest.js
@@ -1,17 +1,0 @@
-/* 
-    Test Entry point for Login Screen
-    version: 2021-11-20
-*/
-
-import React from 'react';
-import LoginScreen from './app/screens/LoginScreen';
-
-
-const App = () => {
-    return (
-        <LoginScreen />
-    );
-
-};
-
-export default App;
